pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "google-genai>=1.29.0",
    "pillow>=11.3.0",
    "psutil>=7.0.0",
    "python-telegram-bot==22.3",
    "sift-stack-py>=0.8.2",
    "telegram>=0.0.1",
    "werkzeug>=3.1.3",
]
